+++ CHANGELOG.md
@@ -29,6 +29,31 @@ You can check your current version with the following command:
     ```
 
 For more information, see [UP42 Python package description](https://pypi.org/project/up42-py/).
+## 1.0.4a11
+
+**Jun 10, 2024**
+
+- Added job templates for `SpaceptAugmentation`, `NSUpsamling`, `Pansharpening` to `templates.py`
+
+## 1.0.4a10
+
+**Jun 7, 2024**
+
+- Added simple single and multi item processing job templates to `templates.py`
+
+## 1.0.4a9
+
+**Jun 7, 2024**
+
+- Added cost evaluation to `JobTemplate` class with number comparison support.
+
+## 1.0.4a8
+
+**Jun 6, 2024**
+
+- Added module `processing.py` with base `JobTemplate` class with post-construct inputs validation.
+- Added helper `SingleItemJobTemplate` and `MultiItemJobTemplate` classes as bases for future specific processing templates.
+
 ## 1.0.4a7
 
 **May 30, 2024**